Adopt paz.transformers in Gemma4: rotary, RMSNorm, logit transforms
Move shared pieces into the library and point Gemma4 at them, with no behavior
change (gemma4 decode / model / sampling / inference tests stay identical):
- RMSNormalization, MergeDims, SplitDim -> paz.layers (de-prefixed from
  Gemma4RMSNormalization; gemma4 attention's GQA reshapes now use the shared
  layers).
- Rotary embedding -> paz.transformers.embeddings.rotary (apply / apply_partial
  / build / build_frequencies), ported from gemma4 layers/core.py.
- gemma4 sampling.py reuses paz.transformers.logits temperature/top-k/top-p.

Co-located unit tests for the new library pieces.

Deferred to a follow-up: KV-cache update reconciliation (Gemma splits K/V while
the shared cache.update stacks them), stateless decode helpers ->
paz.backend.standard, feedforward.glu adoption, reshape_query, and decode-loop
unification.
