+++ examples/gemma4/layers/__init__.py
@@ -1,6 +1,5 @@
 from .attention import attend
 from .decoder import decoder_block
-from .normalization import Gemma4RMSNormalization
 from .normalization import Gemma4VNorm
 from .normalization import ScalarMultiply
 from .normalization import build_rms_norm
