+++ examples/gemma4/layers/core.py
@@ -1,67 +1,5 @@
 import keras
 from keras import ops
-from keras.layers import Layer
-
-
-@keras.saving.register_keras_serializable(package="gemma4")
-class MergeDims(Layer):
-    """Merges two adjacent dims at call time (no parameters)."""
-
-    def __init__(self, axis=-2, **kwargs):
-        super().__init__(**kwargs)
-        self.axis = axis
-
-    def get_config(self):
-        config = super().get_config()
-        config["axis"] = self.axis
-        return config
-
-    def call(self, x):
-        shape = ops.shape(x)
-        idx = self.axis
-        if idx < 0:
-            idx = len(x.shape) + idx
-        merged = shape[idx] * shape[idx + 1]
-        new_shape = shape[:idx] + (merged,) + shape[idx + 2:]
-        return ops.reshape(x, new_shape)
-
-    def compute_output_shape(self, input_shape):
-        idx = self.axis
-        if idx < 0:
-            idx = len(input_shape) + idx
-        a, b = input_shape[idx], input_shape[idx + 1]
-        merged = None if (a is None or b is None) else a * b
-        return input_shape[:idx] + (merged,) + input_shape[idx + 2:]
-
-
-@keras.saving.register_keras_serializable(package="gemma4")
-class SplitDim(Layer):
-    """Splits one dim into two adjacent dims at call time."""
-
-    def __init__(self, axis, sizes, **kwargs):
-        super().__init__(**kwargs)
-        self.axis = axis
-        self.sizes = tuple(sizes)
-
-    def get_config(self):
-        config = super().get_config()
-        config["axis"] = self.axis
-        config["sizes"] = self.sizes
-        return config
-
-    def call(self, x):
-        shape = ops.shape(x)
-        idx = self.axis
-        if idx < 0:
-            idx = len(x.shape) + idx
-        new_shape = shape[:idx] + self.sizes + shape[idx + 1:]
-        return ops.reshape(x, new_shape)
-
-    def compute_output_shape(self, input_shape):
-        idx = self.axis
-        if idx < 0:
-            idx = len(input_shape) + idx
-        return input_shape[:idx] + self.sizes + input_shape[idx + 1:]
 
 
 def build_attention_mask(
@@ -110,82 +48,6 @@ def merge_padding_mask(padding_mask):
         return None
     mask = ops.cast(padding_mask, "bool")
     return ops.expand_dims(mask, axis=1)
-
-
-def apply_partial_rotary_embedding(
-        inputs, wavelength, scaling_factor,
-        partial_rotary_factor, positions=None):
-    head_dim = inputs.shape[-1]
-    raw_dim = int(head_dim * partial_rotary_factor)
-    rotary_dim = max(2, raw_dim - raw_dim % 2)
-    if rotary_dim >= head_dim:
-        return apply_rotary_embedding(
-            inputs, wavelength, scaling_factor, head_dim, positions)
-    half_rotary = rotary_dim // 2
-    half_head = head_dim // 2
-    first_half = inputs[..., :half_head]
-    second_half = inputs[..., half_head:]
-    first_rotary = first_half[..., :half_rotary]
-    second_rotary = second_half[..., :half_rotary]
-    first_static = first_half[..., half_rotary:]
-    second_static = second_half[..., half_rotary:]
-    rotary = ops.concatenate((first_rotary, second_rotary), axis=-1)
-    rotary = apply_rotary_embedding(
-        rotary, wavelength, scaling_factor, head_dim, positions)
-    first_rotary, second_rotary = ops.split(rotary, 2, axis=-1)
-    first_half = ops.concatenate(
-        (first_rotary, first_static), axis=-1)
-    second_half = ops.concatenate(
-        (second_rotary, second_static), axis=-1)
-    return ops.concatenate((first_half, second_half), axis=-1)
-
-
-def apply_rotary_embedding(
-        inputs, wavelength, scaling_factor, denominator, positions=None):
-    cosine, sine = build_rotary_embedding(
-        inputs, wavelength, scaling_factor, denominator, positions)
-    first_half, second_half = ops.split(inputs, 2, axis=-1)
-    rotated = ops.stack((-second_half, first_half), axis=-2)
-    rotated = MergeDims(axis=-2)(rotated)
-    return (inputs * cosine) + (rotated * sine)
-
-
-def build_rotary_embedding(
-        inputs, wavelength, scaling_factor, denominator, positions=None):
-    rotary_dim = inputs.shape[-1]
-    args = (rotary_dim, denominator, wavelength, scaling_factor)
-    inverse = build_inverse_frequencies(*args)
-    if positions is None:
-        positions = build_rotary_positions(inputs)
-    angles = positions * inverse
-    angles = ops.tile(angles, (1, 2))
-    angles = ops.expand_dims(angles, axis=0)
-    if len(inputs.shape) == 4:
-        angles = ops.expand_dims(angles, axis=2)
-    cosine = ops.cast(ops.cos(angles), inputs.dtype)
-    sine = ops.cast(ops.sin(angles), inputs.dtype)
-    return cosine, sine
-
-
-def build_rotary_positions(inputs):
-    trailing = tuple(range(2, len(inputs.shape)))
-    ones = ops.ones_like(inputs)
-    if trailing:
-        ones = ops.mean(ones, axis=trailing)
-    ones = ops.mean(ones, axis=0, keepdims=False)
-    positions = ops.cumsum(ones) - 1.0
-    return ops.expand_dims(positions, axis=1)
-
-
-def build_inverse_frequencies(rotary_dim, denominator, wavelength,
-                              scaling_factor):
-    indices = ops.arange(0, rotary_dim, 2, dtype="float32")
-    denominator = ops.cast(denominator, "float32")
-    frequency = indices / denominator
-    inverse = ops.power(
-        ops.cast(wavelength, "float32"), -frequency)
-    scale = ops.cast(scaling_factor, "float32")
-    return inverse / scale
 
 
 def apply_tanh_soft_cap(values, soft_cap):
