+++ paz/layers/rms_normalization.py
@@ -1,0 +1,34 @@
+import keras
+from keras import ops
+
+
+@keras.saving.register_keras_serializable("layers")
+class RMSNormalization(keras.layers.Layer):
+    def __init__(self, epsilon=1e-6, **kwargs):
+        super().__init__(**kwargs)
+        self.epsilon = epsilon
+
+    def get_config(self):
+        config = super().get_config()
+        config["epsilon"] = self.epsilon
+        return config
+
+    def build(self, input_shape):
+        self.scale = self.add_weight(
+            name="scale",
+            shape=(input_shape[-1],),
+            initializer="ones",
+            trainable=True,
+        )
+        self.built = True
+
+    def compute_output_shape(self, input_shape):
+        return input_shape
+
+    def call(self, x):
+        x = ops.cast(x, "float32")
+        scale = ops.cast(self.scale, "float32")
+        variance = ops.mean(ops.square(x), axis=-1, keepdims=True)
+        output = x * ops.power(variance + self.epsilon, -0.5)
+        output = output * scale
+        return ops.cast(output, self.compute_dtype)
