+++ paz/models/transformers/__init__.py
@@ -10,3 +10,4 @@ from paz.models.transformers import mask
 from paz.models.transformers import logits
 from paz.models.transformers import search
 from paz.models.transformers import feedforward
+from paz.models.transformers import embeddings
