+++ paz/models/transformers/embeddings/__init__.py
@@ -1,0 +1,1 @@
+from paz.models.transformers.embeddings import rotary
